+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 12.4.1
 -----------------
 - Fix PHP warning: fatal error when end date < start date in AddDropBreakdownTime.php & CategoryBreakdownTime.php
 - Fix search discipline numeric field for "0" value in Widget.php
+- SQL case-insensitive username uniqueness check in Student.php, User.php & ImportUsers.fnc.php
 
 Changes in 12.4
 ---------------